successfull callback from custom component through iframe to host panel. breakout iframe content into file
